Moved utils/ApolloClient -> helpers/ApolloClientHelper
