+++ src/services/nodeService.ts
@@ -2,11 +2,11 @@ import { ApolloQueryResult } from 'apollo-client';
 import { LatLng } from 'leaflet';
 import EndpointPath from '~/enums/endpointPath';
 import NodeFactory from '~/factories/nodeFactory';
+import ApolloClient from '~/helpers/ApolloClientHelper';
 import { ILink, INode } from '~/models';
 import { INodeBase, INodeMapHighlight, INodePrimaryKey } from '~/models/INode';
 import IExternalNode from '~/models/externals/IExternalNode';
 import ApiClient from '~/utils/ApiClient';
-import ApolloClient from '~/utils/ApolloClient';
 import GraphqlQueries from './graphqlQueries';
 
 class NodeService {
