+++ src/services/routePathSegmentService.ts
@@ -1,7 +1,7 @@
 import { ApolloQueryResult } from 'apollo-client';
 import RoutePathCopySegmentFactory from '~/factories/routePathCopySegmentFactory';
+import ApolloClient from '~/helpers/ApolloClientHelper';
 import { ICopySegmentRoutePath } from '~/stores/routePathCopySegmentStore';
-import ApolloClient from '~/utils/ApolloClient';
 import GraphqlQueries from './graphqlQueries';
 
 class RoutePathSegmentService {
