+++ src/services/routeService.ts
@@ -2,11 +2,11 @@ import { ApolloQueryResult } from 'apollo-client';
 import EndpointPath from '~/enums/endpointPath';
 import RouteFactory from '~/factories/routeFactory';
 import RoutePathFactory from '~/factories/routePathFactory';
+import ApolloClient from '~/helpers/ApolloClientHelper';
 import { IRoute, IRoutePath } from '~/models';
 import { IRoutePrimaryKey } from '~/models/IRoute';
 import IExternalRoutePath from '~/models/externals/IExternalRoutePath';
 import ApiClient from '~/utils/ApiClient';
-import ApolloClient from '~/utils/ApolloClient';
 import GraphqlQueries from './graphqlQueries';
 
 interface IAllRoutesQueryResult {
