+++ src/services/stopService.ts
@@ -1,9 +1,9 @@
 import { ApolloQueryResult } from 'apollo-client';
 import _ from 'lodash';
+import ApolloClient from '~/helpers/ApolloClientHelper';
 import { IStopItem } from '~/models/IStop';
 import IExternalNode from '~/models/externals/IExternalNode';
 import { IExternalStopItem } from '~/models/externals/IExternalStop';
-import ApolloClient from '~/utils/ApolloClient';
 import GraphqlQueries from './graphqlQueries';
 import NodeService from './nodeService';
