+++ src/utils/ApiClient.ts
@@ -3,10 +3,10 @@ import httpStatusDescriptionCodeList from '~/codeLists/httpStatusDescriptionCode
 import constants from '~/constants/constants';
 import EndpointPath from '~/enums/endpointPath';
 import FetchStatusCode from '~/enums/fetchStatusCode';
+import ApolloClient from '~/helpers/ApolloClientHelper';
 import IError from '~/models/IError';
 import AlertStore from '~/stores/alertStore';
 import LoginStore from '~/stores/loginStore';
-import ApolloClient from '~/utils/ApolloClient';
 
 enum RequestMethod {
     GET = 'GET',
